requirements.txt
numpy
pandas
scikit-learn
matplotlib
Pillow
opencv-python
tqdm
tensorboard
torch
torchvision
